Refactor dirty checking into hook
